refactor(server): Phase 4B extract eval-runner judge transport + share pool
Sub-PR B of Phase 4. Behavior-preserving extraction (no logic changes).

4.1 (partial): pull the self-contained pieces out of the 1870-line
eval-runner.ts into eval-runners/ modules, re-exported from eval-runner.ts so
every import site and test keeps working unchanged:
- eval-runners/judge-calls.ts: judgeComplete + callJudge + callPairwiseJudge +
  callTrajectoryJudge + JudgeOutcome/PairwiseOutcome/TrajectoryOutcome
  (the 6-provider judge transport).
- eval-runners/generate.ts: generateForItem + EvalProvider.
- eval-runners/pool.ts: the async concurrency pool.
eval-runner.ts drops 1870 -> 1251 lines. The deep runEvalRun branch split
(run-trajectory / run-pairwise / run-single) is DEFERRED: it is the
highest-risk, lowest-value part (intricate DB writes, span emission, score
persistence) and getting under 800 lines needs it.

4.2 (partial): experiment-runner.ts now imports the shared pool instead of its
byte-identical copy. Consolidating its local judge logic (buildJudgePrompt /
parseJudgeReply / callJudge) onto the shared modules is DEFERRED on purpose:
the shared judge prompt wording differs, so swapping it would change the scores
this A/B feature produces. That is a product decision, not a refactor.

Verified: server typecheck + lint + test (1316; 107 eval-suite tests green).
